Add Database with tests
The database consists of the methods
- Load   the database from a file (JSON-format is used)
- Save   the database to a file (use JSON-format)
- Get    the value of a key or empty warning with error if nonexisting
- Add    a new key-value-pair to the database

The values are protexted by a Read/Write-Mutex and are kept in memory
during operation unless you explicitly persist them or load a new db
file.

The tests actually write and read to the OS' filesystem. The default
path for the file is the "test.db" in the project's root directory but
this can be overwritten by changing a constant at the top of the
database_test.go file.

Additionally:
- Add test.db to gitignore
- Add a flag to specify the database file on startup
  (not used yet, however)
- Define an interface ShortlinkDB for the above mentioned methods

Signed-off-by: Sascha Scherrer <dev@saschascherrer.de>